+++ src/features/spawn-indicators/Feature.cpp
@@ -7,6 +7,7 @@ using namespace geode::prelude;
 
 void SpawnIndicators::create(EffectGameObject* trigger) {
     if (!trigger->m_isSpawnTriggered) return;
+    if (Cache::layerAlphaMultiplier != 1.0f) return;
     auto pos = Utils::getTriggerBodyPos(trigger);
     Cache::objectDraw->drawCircle(
         pos, // idc if i get scale twice fuck you if you lag cuz of this shit u got bigger problems